+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 ## [0.0.4] - 2024-10-26
 - Initial release of the simple flutter stepper widget.
+
+## [0.0.5] - 2024-10-27
+- add appbar to scaffold for use it.
+- add simple docs for package.